requirements.txt.txt
# LLM & API
openai==2.6.1
python-dotenv==1.1.1

# LangChain & LangGraph
langchain==1.0.5
langchain-core==1.1.0
langchain-community==0.4.1
langgraph==1.0.2
langgraph-checkpoint==3.0.1

# Vector DB (RAG)
chromadb==1.3.4

# Web Search
requests==2.32.5

# UI
gradio==6.0.1

# Utilities
numpy==2.3.5
pydantic==2.12.4